requirements.txt
aws-cdk-lib==2.233.0
constructs>=10.0.0,<11.0.0
cdklabs.generative_ai_cdk_constructs==0.1.312
cdk-nag